requirements.txt
os
datetime
time
streamlit
langchain
langchain-community
langchain-groq
google-generativeai
Pillow
python-dotenv
faiss-cpu